Remove old Tele
